+++ SwitchSnake/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
