+++ SwitchSnake/Resources/Assets.xcassets/AppIcon.appiconset/Icon-1024.png
