+++ SwitchSnake/Resources/Assets.xcassets/Contents.json
