+++ SwitchSnake/Source/Common/Constants.swift
@@ -1,0 +1,15 @@
+//
+//  Constants.swift
+//  SwitchSnake
+//
+//  Created by Mike Price on 08.04.2021.
+//
+
+import Foundation
+
+struct Constants {
+	static let TILE_HEIGHT = 20
+	static let TILE_WIDTH = 15
+	static let START_SNAKE_LENGTH = 10
+	static let FPS = 60.0
+}
