+++ SwitchSnake/Source/Model/MoveDirection.swift
@@ -1,0 +1,15 @@
+//
+//  MoveDirection.swift
+//  SwitchSnake
+//
+//  Created by Mike Price on 08.04.2021.
+//
+
+import Foundation
+
+enum MoveDirection {
+	case right
+	case left
+	case up
+	case down
+}
