+++ SwitchSnake/Source/SwitchSnakeApp.swift
@@ -6,12 +6,15 @@
 //
 
 import SwiftUI
+import AppKit
 
 @main
 struct SwitchSnakeApp: App {
+	@NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate: AppDelegate
+
 	var body: some Scene {
 		WindowGroup {
-			GameView()
+			RootView()
 				.onAppear {
 					NSWindow.allowsAutomaticWindowTabbing = false
 					NSApplication.shared.windows.forEach { $0.tabbingMode = .disallowed }
@@ -26,5 +29,21 @@ struct SwitchSnakeApp: App {
 			CommandGroup(replacing: CommandGroupPlacement.toolbar) {}
 			CommandGroup(replacing: CommandGroupPlacement.sidebar) {}
 		}
+	}
+}
+
+class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
+	@Published var isFocused = true
+
+	func applicationDidBecomeActive(_ notification: Notification) {
+		isFocused = true
+	}
+
+	func applicationWillResignActive(_ notification: Notification) {
+		isFocused = false
+	}
+
+	func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
+		true
 	}
 }
